server/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dotenv-python==0.0.1
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.57.2
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
